+++ test/lib/index.js
@@ -1,2 +1,3 @@
 export * from './extract-hook'
 export * from './formatting'
+export * from './setup'
