+++ test/returns/index.js
@@ -1,0 +1,10 @@
+export * from './create'
+export * from './data'
+export * from './error'
+export * from './isLoading'
+export * from './key'
+export * from './load'
+export * from './refresh'
+export * from './remove'
+export * from './replace'
+export * from './update'
